move intParser to validateArgs.js
