+++ tests/main.test.js
@@ -10,16 +10,11 @@ const easyMocks = [
   ['src/commands/submit.js', ['submitAction']],
   ['src/errorHandler.js', ['handleError']],
   ['src/festive.js', ['printFestiveTitle']],
+  ['src/validation/validateArgs.js', ['intParser']],
 ];
 easyMock(easyMocks);
 mockConfig();
 const mockCommander = (() => {
-  class InvalidArgumentError extends Error {
-    constructor(message) {
-      super(message);
-      this.name = 'InvalidArgumentError';
-    }
-  }
   const toReturn = {
     name: jest.fn().mockReturnThis(),
     description: jest.fn().mockReturnThis(),
@@ -43,13 +38,11 @@ const mockCommander = (() => {
         argParser: jest.fn().mockReturnThis(),
       };
     },
-    InvalidArgumentError,
   }));
   return toReturn;
 })();
 
 // import after mocks set up
-const { InvalidArgumentError } = await import('commander');
 const {
   authAction,
   initAction,
@@ -58,7 +51,6 @@ const {
   submitAction,
   handleError,
 } = await easyResolve(easyMocks);
-const { intParser } = await import('../src/main.js');
 
 describe('main', () => {
   beforeEach(() => {
@@ -99,26 +91,4 @@ describe('main', () => {
       await import('../src/main.js');
       expect(handleError).toHaveBeenCalledWith(error);
     }));
-});
-
-describe('argParser()', () => {
-  test.each([null, undefined, '', false, true, {}, Promise.resolve(true)])(
-    'throws if not parsable as int: %s',
-    (value) => {
-      expect(() => {
-        intParser([1, 2, 3])(value);
-      }).toThrow(InvalidArgumentError);
-    }
-  );
-
-  test('throws if value is not a valid choice', () => {
-    expect(() => {
-      intParser([1, 2, 3])('4');
-    }).toThrow(InvalidArgumentError);
-  });
-
-  test('returns integer if value is a valid choice', () => {
-    const result = intParser([1, 2, 3])('1');
-    expect(result).toBe(1);
-  });
 });
